refactor for reusable action buttons
