+++ src/applications/personalization/profile360/components/AddressSection.jsx
@@ -1,25 +1,17 @@
 import React from 'react';
 
-import { kebabCase, toLower } from 'lodash';
+import { kebabCase } from 'lodash';
 
 import HeadingWithEdit from './HeadingWithEdit';
 import Modal from '@department-of-veterans-affairs/formation/Modal';
 import Address from './Address';
 import LoadingButton from './LoadingButton';
 import Transaction from './Transaction';
+import FormActionButtons from './FormActionButtons';
 import AlertBox from '@department-of-veterans-affairs/formation/AlertBox';
 import { consolidateAddress, expandAddress, isEmptyAddress, formatAddress } from '../../../../platform/forms/address/helpers';
 
 class EditAddressModal extends React.Component {
-
-  constructor(props) {
-    super(props);
-
-    this.state = {
-      deleteInitiated: false,
-    };
-  }
-
   componentDidMount() {
     let defaultFieldValue = {};
     if (this.props.addressData) {
@@ -47,46 +39,11 @@ class EditAddressModal extends React.Component {
   }
 
   renderActionButtons() {
-    const cancelDeleteAction = () => {
-      this.setState({ deleteInitiated: false });
-    };
-
-    const confirmDeleteAction = (e) => {
-      e.preventDefault();
-      this.props.onDelete();
-    };
-
-    const alertContent = (
-      <div>
-        <h3>Are you sure?</h3>
-        <p>This will delete your {toLower(this.props.title)} across many VA records. You can always come back to your profile later if you'd like to add this information back in.</p>
-        <div>
-          <LoadingButton isLoading={this.props.isLoading} onClick={confirmDeleteAction}>Confirm</LoadingButton>
-          <button type="button" className="usa-button-secondary" onClick={cancelDeleteAction}>Cancel</button>
-        </div>
-      </div>
-    );
-
-    if (this.state.deleteInitiated) {
-      return (
-        <AlertBox
-          isVisible
-          status="warning"
-          content={alertContent}/>
-      );
-    }
-
     return (
-      <div>
+      <FormActionButtons onCancel={this.props.onCancel} onDelete={this.props.onDelete}>
         <LoadingButton isLoading={this.props.isLoading}>Update</LoadingButton>
         <button type="button" className="usa-button-secondary" onClick={this.props.onCancel}>Cancel</button>
-        <div className="right">
-          <button className="usa-button-secondary button-link"
-            onClick={() => this.setState({ deleteInitiated: true })}>
-            <i className="fa fa-trash"></i> <span>Delete</span>
-          </button>
-        </div>
-      </div>
+      </FormActionButtons>
     );
   }
